Undo renaming of Mobile BE file
