+++ deploy/scripts/src/mobile/v2-mobile-backend-get-client-attestation.ts
